Retrieve site mementos from the Internet Archive

Mementos are prior versions of web pages that have been cached from web crawlers. They can be found in web archives (such as the Internet Archive) or systems that support versioning such as wikis or revision control systems.

get_mementos(url, timestamp = format(Sys.Date(), "%Y"))

Arguments

url
URL to retrieve information for

timestamp
(optional) timestamp to use when checking for availability. If not specified, the most recenty available capture in Wayback is returned. If you don't pass in a valid R "time-y" object, you will need to ensure the character string you provide is in a valid subset of `YYYYMMDDhhmmss`.

Details

It's possible that a memento list could be large and contain an entry for the next "page" of mementos.

TODO Not sure we should handle that ^^ here since this is a free service and is already pretty taxed. If the user needs it they can request from a new timestamp. But I'm not committed to this concept.

Examples

# NOT RUN {
rproj_mnto <- get_mementos("https://www.r-project.org/")

dplyr::glimpse(rproj_mnto)
## Observations: 7
## Variables: 3
## $ link <chr> "http://www.r-project.org/", "http://web.archive.org/web/timemap/...
## $ rel  <chr> "original", "timemap", "timegate", "first memento", "prev memento...
## $ ts   <dttm> NA, NA, NA, 2000-06-20 19:56:31, 2017-08-29 04:41:15, 2017-08-30...
# }